requirements.txt
pandas
scikit-learn
mysql-connector
requests
matplotlib
seaborn